Exercise 2 Transaction Statistics Tests

+++ challenge/src/main/java/com/ceviant/coding/challenge/exercise1/service/TransactionServiceImpl.java
@@ -13,7 +13,7 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Override
     public void addTransaction(Transaction transaction){
-        statistics.addNewTransaction(transaction.getAmount().doubleValue());
+        statistics.addNewTransaction(transaction.getAmount());
     }
 
     @Override
